[#32] refactor: useCommand BlockContent 로직에 적용
